Move grouped GEMM backend harnesses out of example
stack-info: PR: https://github.com/pytorch/helion/pull/3038, branch: choijon5/grouped-gemm-cute-stack/5
