+++ test/test_examples.py
@@ -1,9 +1,6 @@
 from __future__ import annotations
 
-import ast
-from itertools import starmap
 import math
-import os
 from typing import Any
 import unittest
 from unittest.mock import patch
@@ -42,7 +39,6 @@ from helion._testing import skipIfXPU
 from helion._testing import xfailIfPallas
 from helion._testing import xfailIfPallasInterpret
 from helion._testing import xfailIfPallasTpu
-import helion.language as hl
 from helion.runtime.config import Config
 from helion.runtime.ref_mode import is_ref_mode_enabled
 
@@ -71,240 +67,6 @@ def _compile_only(
         bound._config = config
         return bound
     return bound.compile_config(config)
-
-
-def _cute_wrapper_plans_from_code(code: str) -> list[dict[str, object]]:
-    marker = "._helion_cute_wrapper_plans = "
-    line = next(line for line in code.splitlines() if marker in line)
-    payload = line.split(marker, 1)[1]
-    freeze_prefix = "helion.runtime._freeze_cute_wrapper_plans("
-    if payload.startswith(freeze_prefix) and payload.endswith(")"):
-        payload = payload[len(freeze_prefix) : -1]
-    return list(ast.literal_eval(payload))
-
-
-def _deepgemm_segment_test_kernel(body: object) -> helion.Kernel:
-    return helion.kernel(
-        static_shapes=False,
-        config=helion.Config(block_sizes=[64, 64, 64]),
-    )(body)
-
-
-def _deepgemm_segment_bad_store_row_body(
-    a_packed: torch.Tensor,
-    b_grouped: torch.Tensor,
-    work_tile_metadata: torch.Tensor,
-) -> torch.Tensor:
-    m_total_aligned, k = a_packed.shape
-    _g, n, k2 = b_grouped.shape
-    assert k == k2
-    assert work_tile_metadata.size(1) == 4
-    block_m = hl.register_block_size(64)
-    block_n = hl.register_block_size(64)
-    block_k = hl.register_block_size(64)
-    out = torch.zeros(
-        m_total_aligned,
-        n,
-        dtype=a_packed.dtype,
-        device=a_packed.device,
-    )
-    for work_tile, tile_m, tile_n in hl.tile(
-        [work_tile_metadata.size(0), 64, n],
-        block_size=[1, block_m, block_n],
-    ):
-        work_id = work_tile.begin
-        group_id = work_tile_metadata[work_id, 0]
-        global_m_start = work_tile_metadata[work_id, 1]
-        valid_m = work_tile_metadata[work_id, 2]
-        local_m = tile_m.index
-        row_index = global_m_start + local_m
-        valid_rows = local_m < valid_m
-        acc = hl.zeros([tile_m, tile_n], dtype=torch.float32)
-        for tile_k in hl.tile(k, block_size=block_k):
-            a_blk = hl.load(
-                a_packed,
-                [row_index, tile_k],
-                extra_mask=valid_rows[:, None],  # pyrefly: ignore[bad-index]
-            )
-            acc = torch.addmm(
-                acc,
-                a_blk,
-                b_grouped[group_id, tile_n, tile_k].T,
-            )
-        hl.store(
-            out,
-            [row_index + 1, tile_n],
-            acc.to(out.dtype),
-            extra_mask=valid_rows[:, None],  # pyrefly: ignore[bad-index]
-        )
-    return out
-
-
-def _deepgemm_segment_bad_store_mask_body(
-    a_packed: torch.Tensor,
-    b_grouped: torch.Tensor,
-    work_tile_metadata: torch.Tensor,
-) -> torch.Tensor:
-    m_total_aligned, k = a_packed.shape
-    _g, n, k2 = b_grouped.shape
-    assert k == k2
-    assert work_tile_metadata.size(1) == 4
-    block_m = hl.register_block_size(64)
-    block_n = hl.register_block_size(64)
-    block_k = hl.register_block_size(64)
-    out = torch.zeros(
-        m_total_aligned,
-        n,
-        dtype=a_packed.dtype,
-        device=a_packed.device,
-    )
-    for work_tile, tile_m, tile_n in hl.tile(
-        [work_tile_metadata.size(0), 64, n],
-        block_size=[1, block_m, block_n],
-    ):
-        work_id = work_tile.begin
-        group_id = work_tile_metadata[work_id, 0]
-        global_m_start = work_tile_metadata[work_id, 1]
-        valid_m = work_tile_metadata[work_id, 2]
-        local_m = tile_m.index
-        row_index = global_m_start + local_m
-        valid_rows = local_m < valid_m
-        store_rows = valid_rows & (local_m >= 0)
-        acc = hl.zeros([tile_m, tile_n], dtype=torch.float32)
-        for tile_k in hl.tile(k, block_size=block_k):
-            a_blk = hl.load(
-                a_packed,
-                [row_index, tile_k],
-                extra_mask=valid_rows[:, None],  # pyrefly: ignore[bad-index]
-            )
-            acc = torch.addmm(
-                acc,
-                a_blk,
-                b_grouped[group_id, tile_n, tile_k].T,
-            )
-        hl.store(
-            out,
-            [row_index, tile_n],
-            acc.to(out.dtype),
-            extra_mask=store_rows[:, None],  # pyrefly: ignore[bad-index]
-        )
-    return out
-
-
-def _deepgemm_segment_store_m_mask_body(
-    a_packed: torch.Tensor,
-    b_grouped: torch.Tensor,
-    work_tile_metadata: torch.Tensor,
-) -> torch.Tensor:
-    m_total_aligned, k = a_packed.shape
-    _g, n, k2 = b_grouped.shape
-    assert k == k2
-    assert work_tile_metadata.size(1) == 4
-    block_m = hl.register_block_size(64)
-    block_n = hl.register_block_size(64)
-    block_k = hl.register_block_size(64)
-    out = torch.zeros(
-        m_total_aligned,
-        n,
-        dtype=a_packed.dtype,
-        device=a_packed.device,
-    )
-    for work_tile, tile_m, tile_n in hl.tile(
-        [work_tile_metadata.size(0), 64, n],
-        block_size=[1, block_m, block_n],
-    ):
-        work_id = work_tile.begin
-        group_id = work_tile_metadata[work_id, 0]
-        global_m_start = work_tile_metadata[work_id, 1]
-        valid_m = work_tile_metadata[work_id, 2]
-        store_m = work_tile_metadata[work_id, 3]
-        local_m = tile_m.index
-        row_index = global_m_start + local_m
-        valid_rows = local_m < valid_m
-        store_rows = local_m < store_m
-        acc = hl.zeros([tile_m, tile_n], dtype=torch.float32)
-        for tile_k in hl.tile(k, block_size=block_k):
-            a_blk = hl.load(
-                a_packed,
-                [row_index, tile_k],
-                extra_mask=valid_rows[:, None],  # pyrefly: ignore[bad-index]
-            )
-            acc = torch.addmm(
-                acc,
-                a_blk,
-                b_grouped[group_id, tile_n, tile_k].T,
-            )
-        hl.store(
-            out,
-            [row_index, tile_n],
-            acc.to(out.dtype),
-            extra_mask=store_rows[:, None],  # pyrefly: ignore[bad-index]
-        )
-    return out
-
-
-def _deepgemm_segment_inverted_row_mask_body(
-    a_packed: torch.Tensor,
-    b_grouped: torch.Tensor,
-    work_tile_metadata: torch.Tensor,
-) -> torch.Tensor:
-    m_total_aligned, k = a_packed.shape
-    _g, n, k2 = b_grouped.shape
-    assert k == k2
-    assert work_tile_metadata.size(1) == 4
-    block_m = hl.register_block_size(64)
-    block_n = hl.register_block_size(64)
-    block_k = hl.register_block_size(64)
-    out = torch.zeros(
-        m_total_aligned,
-        n,
-        dtype=a_packed.dtype,
-        device=a_packed.device,
-    )
-    for work_tile, tile_m, tile_n in hl.tile(
-        [work_tile_metadata.size(0), 64, n],
-        block_size=[1, block_m, block_n],
-    ):
-        work_id = work_tile.begin
-        group_id = work_tile_metadata[work_id, 0]
-        global_m_start = work_tile_metadata[work_id, 1]
-        valid_m = work_tile_metadata[work_id, 2]
-        local_m = tile_m.index
-        row_index = global_m_start + local_m
-        valid_rows = valid_m < local_m
-        acc = hl.zeros([tile_m, tile_n], dtype=torch.float32)
-        for tile_k in hl.tile(k, block_size=block_k):
-            a_blk = hl.load(
-                a_packed,
-                [row_index, tile_k],
-                extra_mask=valid_rows[:, None],  # pyrefly: ignore[bad-index]
-            )
-            acc = torch.addmm(
-                acc,
-                a_blk,
-                b_grouped[group_id, tile_n, tile_k].T,
-            )
-        hl.store(
-            out,
-            [row_index, tile_n],
-            acc.to(out.dtype),
-            extra_mask=valid_rows[:, None],  # pyrefly: ignore[bad-index]
-        )
-    return out
-
-
-_deepgemm_segment_bad_store_row_kernel = _deepgemm_segment_test_kernel(
-    _deepgemm_segment_bad_store_row_body
-)
-_deepgemm_segment_bad_store_mask_kernel = _deepgemm_segment_test_kernel(
-    _deepgemm_segment_bad_store_mask_body
-)
-_deepgemm_segment_store_m_mask_kernel = _deepgemm_segment_test_kernel(
-    _deepgemm_segment_store_m_mask_body
-)
-_deepgemm_segment_inverted_row_mask_kernel = _deepgemm_segment_test_kernel(
-    _deepgemm_segment_inverted_row_mask_body
-)
 
 
 def setUpModule() -> None:
@@ -1911,7 +1673,7 @@ class TestExamples(RefEagerTestBase, TestCase):
             fn_name="grouped_gemm_jagged",
         )
 
-    @xfailIfPallas("Pallas scatter: multiple indirect dims are not supported")
+    @xfailIfPallas("CUDA-specific code paths")
     def test_grouped_gemm_jagged_persistent(self):
         # Build small jagged grouped GEMM inputs
         torch.manual_seed(0)
@@ -1947,2203 +1709,6 @@ class TestExamples(RefEagerTestBase, TestCase):
             expected,
             fn_name="grouped_gemm_jagged_persistent",
         )
-
-    @onlyBackends(["cute"])
-    def test_grouped_gemm_deepgemm_m_grouped_bf16_nt_contiguous(self):
-        torch.manual_seed(0)
-        mod = import_path(EXAMPLES_DIR / "grouped_gemm.py")
-        use_generated_default = (
-            os.environ.get("HELION_CUTE_MMA_IMPL", "").strip().lower() == "tcgen05"
-        )
-        if use_generated_default:
-            from helion._compat import requires_cuda_version
-            from helion._compiler.cute.mma_support import get_cute_mma_support
-
-            if DEVICE.type != "cuda":
-                self.skipTest("CuTeDSL generated segment path requires CUDA")
-            if not requires_cuda_version("13"):
-                self.skipTest("CuTeDSL generated segment path requires CUDA >= 13")
-            with torch.cuda.device(DEVICE):
-                major, _minor = torch.cuda.get_device_capability(DEVICE)
-                if major < 10:
-                    self.skipTest("CuTeDSL generated segment path requires SM100+")
-                if not get_cute_mma_support().tcgen05_f16bf16:
-                    self.skipTest(
-                        "tcgen05 F16/BF16 MMA is not supported on this machine"
-                    )
-
-        args, expected = mod.make_deepgemm_m_grouped_bf16_gemm_nt_contiguous_args(
-            (188, 64, 40, 64) if use_generated_default else (64, 96, 128, 32),
-            n=128 if use_generated_default else 64,
-            k=64,
-            m_alignment=224 if use_generated_default else 64,
-            tail_padding=0 if use_generated_default else 64,
-            device=DEVICE,
-        )
-        layout_has_valid_prefix_tiles = (
-            mod._validate_deepgemm_m_grouped_bf16_gemm_nt_contiguous_args(*args)
-        )
-        if use_generated_default:
-            self.assertFalse(layout_has_valid_prefix_tiles)
-            self.assertTrue(
-                mod._deepgemm_m_grouped_bf16_gemm_nt_contiguous_use_generated_segment_kernel(
-                    *args,
-                    layout_has_valid_prefix_tiles=layout_has_valid_prefix_tiles,
-                )
-            )
-            self.assertTrue(
-                mod._deepgemm_m_grouped_bf16_gemm_nt_contiguous_use_selected_segment_kernel(
-                    *args
-                )
-            )
-
-        actual = mod.deepgemm_m_grouped_bf16_gemm_nt_contiguous(*args)
-
-        torch.testing.assert_close(actual, expected, rtol=2e-2, atol=2e-2)
-        self.assertTrue(bool(torch.all(actual[args[2] == -1] == 0).item()))
-
-        row_hole_layout = args[2].clone()
-        row_hole_layout[16:32] = -1
-        row_hole_layout[32:64] = 0
-        with self.assertRaisesRegex(ValueError, "strictly increasing"):
-            mod._validate_deepgemm_m_grouped_bf16_gemm_nt_contiguous_args(
-                args[0],
-                args[1],
-                row_hole_layout,
-            )
-
-    def _skip_unless_blackwell_grouped_gemm(self) -> None:
-        from helion._compat import requires_cuda_version
-        from helion._compiler.cute.mma_support import get_cute_mma_support
-
-        if DEVICE.type != "cuda":
-            self.skipTest("Blackwell grouped GEMM requires CUDA")
-        if not requires_cuda_version("13"):
-            self.skipTest("Blackwell grouped GEMM requires CUDA >= 13")
-        with torch.cuda.device(DEVICE):
-            major, _minor = torch.cuda.get_device_capability(DEVICE)
-            if major < 10:
-                self.skipTest("Blackwell grouped GEMM requires SM100+")
-            if not get_cute_mma_support().tcgen05_f16bf16:
-                self.skipTest("tcgen05 F16/BF16 MMA is not supported on this machine")
-
-    @onlyBackends(["cute"])
-    @skipIfNotCUDA()
-    def test_grouped_gemm_blackwell_nt_accepts_heterogeneous_fp16_metadata(self):
-        from helion._compiler.cute import grouped_deepgemm
-
-        mod = import_path(EXAMPLES_DIR / "grouped_gemm.py")
-        group_A = tuple(
-            torch.empty(m, k, device=DEVICE, dtype=torch.float16)
-            for m, _n, k, _l in mod.BLACKWELL_GROUPED_GEMM_DOC_PROBLEM_SIZES
-        )
-        group_B = tuple(
-            torch.empty(n, k, device=DEVICE, dtype=torch.float16)
-            for _m, n, k, _l in mod.BLACKWELL_GROUPED_GEMM_DOC_PROBLEM_SIZES
-        )
-        out_groups = tuple(
-            torch.empty(m, n, device=DEVICE, dtype=torch.float16)
-            for m, n, _k, _l in mod.BLACKWELL_GROUPED_GEMM_DOC_PROBLEM_SIZES
-        )
-
-        metadata = grouped_deepgemm._blackwell_grouped_gemm_nt_metadata(
-            group_A,
-            group_B,
-            out_groups,
-        )
-        self.assertEqual(
-            metadata.problems,
-            tuple(
-                starmap(
-                    grouped_deepgemm.BlackwellGroupedGemmProblem,
-                    mod.BLACKWELL_GROUPED_GEMM_DOC_PROBLEM_SIZES,
-                )
-            ),
-        )
-        self.assertEqual(
-            grouped_deepgemm._compute_blackwell_total_clusters(
-                metadata.problems,
-                mma_tiler_mn=(128, 64),
-                cluster_shape_mn=(1, 1),
-                use_2cta_instrs=False,
-            ),
-            sum(
-                math.ceil(m / 128) * math.ceil(n / 64)
-                for m, n, _k, _l in mod.BLACKWELL_GROUPED_GEMM_DOC_PROBLEM_SIZES
-            ),
-        )
-        self.assertFalse(metadata.a_mode0_major)
-        self.assertFalse(metadata.b_mode0_major)
-        self.assertFalse(metadata.c_mode0_major)
-
-        bad_group_A = list(group_A)
-        m0, k0 = int(group_A[0].size(0)), int(group_A[0].size(1))
-        padded = torch.empty(m0, k0 + 1, device=DEVICE, dtype=torch.float16)
-        bad_group_A[0] = padded[:, :k0]
-        self.assertEqual(bad_group_A[0].shape, group_A[0].shape)
-        self.assertEqual(bad_group_A[0].stride(1), 1)
-        leading_stride_bytes = bad_group_A[0].stride(0) * bad_group_A[0].element_size()
-        self.assertNotEqual(leading_stride_bytes % 16, 0)
-        with self.assertRaisesRegex(ValueError, "leading stride.*16-byte"):
-            grouped_deepgemm._blackwell_grouped_gemm_nt_metadata(
-                tuple(bad_group_A),
-                group_B,
-                out_groups,
-            )
-
-    @skipIfRefEager("Argument validation does not execute a Helion kernel")
-    def test_grouped_gemm_blackwell_nt_arg_builder_validates_alignment(self):
-        mod = import_path(EXAMPLES_DIR / "grouped_gemm.py")
-        device = torch.device("cpu")
-
-        self.assertEqual(
-            mod.BLACKWELL_GROUPED_GEMM_DEFAULT_PROBLEM_SIZES,
-            (
-                (128, 128, 128, 1),
-                (512, 128, 128, 1),
-                (128, 256, 128, 1),
-            ),
-        )
-        with self.assertRaisesRegex(ValueError, "K.*16-byte"):
-            mod.make_blackwell_grouped_gemm_nt_args(
-                ((128, 128, 130, 1),),
-                dtype=torch.float16,
-                device=device,
-            )
-        with self.assertRaisesRegex(ValueError, "N.*16-byte"):
-            mod.make_blackwell_grouped_gemm_nt_args(
-                ((128, 130, 128, 1),),
-                dtype=torch.float16,
-                device=device,
-            )
-        padded = torch.empty(128, 129, device=device, dtype=torch.float16)
-        with self.assertRaisesRegex(ValueError, "leading stride.*16-byte"):
-            mod._blackwell_grouped_gemm_nt_generated_check_16b_alignment(
-                padded[:, :128],
-                major="K",
-                mode0="M",
-                tensor_name="A",
-                group=0,
-            )
-
-    @onlyBackends(["cute"])
-    @skipIfNotCUDA()
-    @patch.dict(
-        os.environ,
-        {
-            "HELION_BACKEND": "cute",
-            "HELION_CUTE_MMA_IMPL": "tcgen05",
-            "HELION_AUTOTUNE_EFFORT": "none",
-        },
-        clear=False,
-    )
-    def test_grouped_gemm_blackwell_nt_generated_direct_source(self):
-        self._skip_unless_blackwell_grouped_gemm()
-        torch.manual_seed(0)
-        mod = import_path(EXAMPLES_DIR / "grouped_gemm.py")
-        args, _expected = mod.make_blackwell_grouped_gemm_nt_args(
-            mod.BLACKWELL_GROUPED_GEMM_DOC_PROBLEM_SIZES,
-            dtype=torch.float16,
-            device=DEVICE,
-        )
-        group_A, group_B = args
-        out_groups = tuple(
-            torch.empty(
-                int(a.size(0)),
-                int(b.size(0)),
-                device=DEVICE,
-                dtype=torch.float16,
-            )
-            for a, b in zip(group_A, group_B, strict=True)
-        )
-        prepared = mod._blackwell_grouped_gemm_nt_generated_args(
-            group_A,
-            group_B,
-            out_groups,
-            mma_tiler_mn=(128, 64),
-            cluster_shape_mn=(1, 1),
-        )
-        self.assertIsNotNone(prepared)
-        assert prepared is not None
-        kernel_args, _out_tuple = prepared
-        direct_pointers = kernel_args[-2]
-        self.assertEqual(
-            direct_pointers.detach().cpu().tolist(),
-            [
-                [int(a.data_ptr()), int(b.data_ptr()), int(out.data_ptr())]
-                for a, b, out in zip(group_A, group_B, out_groups, strict=True)
-            ],
-        )
-
-        bound = mod._blackwell_grouped_gemm_nt_generated_kernel.bind(kernel_args)
-        bound.env.config_spec.cute_tcgen05_search_enabled = True
-        code = bound.to_triton_code(
-            mod._blackwell_grouped_gemm_nt_generated_config(reserved_sms=4)
-        )
-        self.assertIn("PipelineTmaUmma.create(num_stages=3", code)
-        self.assertIn("'external_direct_pointer_metadata': True", code)
-        self.assertIn("'tcgen05_grouped_static_reserved_sms': 4", code)
-        self.assertIn("tcgen05_grouped_direct_pointers", code)
-        self.assertIn("tcgen05_grouped_direct_strides", code)
-        self.assertIn("tcgen05_grouped_tensormap_a_addr = (", code)
-        self.assertIn("tcgen05_grouped_tensormap_b_addr = (", code)
-        self.assertIn("tcgen05_grouped_d_tensormap_addr = (", code)
-        self.assertIn("cute.make_ptr(cutlass.Float16", code)
-        self.assertNotIn(
-            "tcgen05_grouped_tensormap_a_base = a_placeholder.iterator +",
-            code,
-        )
-        self.assertNotIn(
-            "tcgen05_grouped_tensormap_b_base = b_placeholder.iterator +",
-            code,
-        )
-        self.assertNotIn(
-            "tcgen05_grouped_d_tensormap_base = out_placeholder.iterator +",
-            code,
-        )
-        self.assertNotIn("_cutlass_grouped_gemm_kernel", code)
-        self.assertNotIn("GroupedGemmKernel", code)
-        self.assertNotIn("torch.cat", code)
-        self.assertNotIn("torch.stack", code)
-
-        default_args, _default_expected = mod.make_blackwell_grouped_gemm_nt_args(
-            mod.BLACKWELL_GROUPED_GEMM_DEFAULT_PROBLEM_SIZES,
-            dtype=torch.float16,
-            device=DEVICE,
-        )
-        default_group_A, default_group_B = default_args
-        default_out_groups = tuple(
-            torch.empty(
-                int(a.size(0)),
-                int(b.size(0)),
-                device=DEVICE,
-                dtype=torch.float16,
-            )
-            for a, b in zip(default_group_A, default_group_B, strict=True)
-        )
-        default_prepared = mod._blackwell_grouped_gemm_nt_generated_args(
-            default_group_A,
-            default_group_B,
-            default_out_groups,
-            mma_tiler_mn=(128, 64),
-            cluster_shape_mn=(1, 1),
-        )
-        self.assertIsNotNone(default_prepared)
-        assert default_prepared is not None
-        default_kernel_args, _default_out_tuple = default_prepared
-        default_bound = mod._blackwell_grouped_gemm_nt_generated_kernel.bind(
-            default_kernel_args
-        )
-        default_bound.env.config_spec.cute_tcgen05_search_enabled = True
-        default_code = default_bound.to_triton_code(
-            mod._blackwell_grouped_gemm_nt_generated_config()
-        )
-        self.assertIn("PipelineTmaUmma.create(num_stages=2", default_code)
-        self.assertNotIn("tcgen05_grouped_static_reserved_sms", default_code)
-
-    @skipIfRefEager("Generated config selection does not execute a Helion kernel")
-    def test_grouped_gemm_blackwell_nt_generated_ab_stage_selector(self):
-        captured_configs: list[Config] = []
-        call_count = 0
-        device = torch.device("cpu")
-
-        class FakeConfigSpec:
-            cute_tcgen05_search_enabled = False
-
-        class FakeEnv:
-            def __init__(self) -> None:
-                self.config_spec = FakeConfigSpec()
-
-        class FakeBound:
-            def __init__(self) -> None:
-                self.env = FakeEnv()
-
-            def set_config(self, config: Config) -> None:
-                captured_configs.append(config)
-
-            def __call__(self, *args: object) -> None:
-                nonlocal call_count
-                call_count += 1
-
-        def fake_bind(args: tuple[torch.Tensor, ...]) -> FakeBound:
-            return FakeBound()
-
-        def run_case(mod: Any, k: int) -> None:
-            group_A = (torch.empty(128, k, device=device, dtype=torch.float16),)
-            group_B = (torch.empty(64, k, device=device, dtype=torch.float16),)
-            out_groups = (torch.empty(128, 64, device=device, dtype=torch.float16),)
-            kernel_args = (torch.empty(1, device=device, dtype=torch.float16),)
-            with patch.object(
-                mod,
-                "_blackwell_grouped_gemm_nt_generated_args",
-                return_value=(kernel_args, out_groups),
-            ) as generated_args:
-                result = mod._blackwell_grouped_gemm_nt_generated(
-                    group_A,
-                    group_B,
-                    out_groups,
-                    mma_tiler_mn=(128, 64),
-                    cluster_shape_mn=(1, 1),
-                    use_2cta_instrs=False,
-                )
-            generated_args.assert_called_once_with(
-                group_A,
-                group_B,
-                out_groups,
-                mma_tiler_mn=(128, 64),
-                cluster_shape_mn=(1, 1),
-                use_2cta_instrs=False,
-            )
-            self.assertIs(result, out_groups)
-
-        with patch.dict(
-            os.environ,
-            {
-                "HELION_BACKEND": "cute",
-                "HELION_CUTE_MMA_IMPL": "tcgen05",
-                "HELION_AUTOTUNE_EFFORT": "none",
-            },
-            clear=False,
-        ):
-            mod: Any = import_path(EXAMPLES_DIR / "grouped_gemm.py")
-            with (
-                patch.object(
-                    mod._blackwell_grouped_gemm_nt_generated_kernel,
-                    "bind",
-                    side_effect=fake_bind,
-                ),
-                patch.object(
-                    mod,
-                    "_blackwell_grouped_gemm_nt_generated_reserved_sms",
-                    return_value=0,
-                ),
-            ):
-                run_case(mod, 128)
-                run_case(mod, 192)
-
-        self.assertEqual(call_count, 2)
-        self.assertEqual(len(captured_configs), 2)
-        small_k_config = captured_configs[0].config
-        large_k_config = captured_configs[1].config
-        self.assertEqual(small_k_config["block_sizes"], [128, 64, 64])
-        self.assertEqual(small_k_config["num_stages"], 2)
-        self.assertNotIn("tcgen05_ab_stages", small_k_config)
-        self.assertEqual(large_k_config["block_sizes"], [128, 64, 64])
-        self.assertEqual(large_k_config["tcgen05_ab_stages"], 4)
-
-    @onlyBackends(["cute"])
-    def test_grouped_gemm_blackwell_nt_generated_reserved_sms_selector(self):
-        mod = import_path(EXAMPLES_DIR / "grouped_gemm.py")
-        problem = mod._BlackwellGeneratedGemmProblem
-        self.assertEqual(
-            mod._blackwell_grouped_gemm_nt_generated_problem_reserved_sms(
-                (
-                    problem(8192, 1280, 32),
-                    problem(16, 384, 1536),
-                    problem(640, 1280, 16),
-                    problem(640, 160, 16),
-                ),
-                num_sm=148,
-                block_m=128,
-                block_n=64,
-            ),
-            3,
-        )
-        self.assertEqual(
-            mod._blackwell_grouped_gemm_nt_generated_problem_reserved_sms(
-                (
-                    problem(8192, 1280, 32),
-                    problem(128, 384, 1536),
-                    problem(640, 1280, 16),
-                    problem(640, 192, 16),
-                ),
-                num_sm=148,
-                block_m=128,
-                block_n=64,
-            ),
-            0,
-        )
-        self.assertEqual(
-            mod._blackwell_grouped_gemm_nt_generated_problem_reserved_sms(
-                (
-                    problem(8192, 1280, 32),
-                    problem(16, 384, 1536),
-                    problem(640, 1280, 16),
-                    problem(640, 192, 16),
-                ),
-                num_sm=148,
-                block_m=128,
-                block_n=64,
-            ),
-            3,
-        )
-        self.assertEqual(
-            mod._blackwell_grouped_gemm_nt_generated_problem_reserved_sms(
-                (
-                    problem(8192, 1280, 32),
-                    problem(16, 384, 1536),
-                    problem(640, 1280, 16),
-                    problem(640, 128, 16),
-                ),
-                num_sm=148,
-                block_m=128,
-                block_n=64,
-            ),
-            0,
-        )
-        self.assertEqual(
-            mod._blackwell_grouped_gemm_nt_generated_problem_reserved_sms(
-                (
-                    problem(8065, 1280, 32),
-                    problem(128, 384, 1536),
-                    problem(640, 1280, 16),
-                    problem(640, 128, 16),
-                ),
-                num_sm=148,
-                block_m=128,
-                block_n=64,
-            ),
-            2,
-        )
-        self.assertNotIn(
-            "tcgen05_grouped_static_reserved_sms",
-            mod._blackwell_grouped_gemm_nt_generated_config().config,
-        )
-
-    @onlyBackends(["cute"])
-    @skipIfNotCUDA()
-    @patch.dict(
-        os.environ,
-        {
-            "HELION_BACKEND": "cute",
-            "HELION_CUTE_MMA_IMPL": "tcgen05",
-            "HELION_AUTOTUNE_EFFORT": "none",
-        },
-        clear=False,
-    )
-    def test_grouped_gemm_blackwell_nt_generated_declines_common_envelope(self):
-        self._skip_unless_blackwell_grouped_gemm()
-        from helion._compiler.cute import grouped_deepgemm
-
-        grouped_deepgemm._BLACKWELL_GROUPED_COMPILED_CACHE.clear()
-        grouped_deepgemm._BLACKWELL_GROUPED_PREPARED_CACHE.clear()
-        torch.manual_seed(0)
-        mod = import_path(EXAMPLES_DIR / "grouped_gemm.py")
-        args, _expected = mod.make_blackwell_grouped_gemm_nt_args(
-            ((128, 160, 80, 1),),
-            dtype=torch.float16,
-            device=DEVICE,
-        )
-        group_A, group_B = args
-        out_groups = tuple(
-            torch.empty(
-                int(a.size(0)),
-                int(b.size(0)),
-                device=DEVICE,
-                dtype=torch.float16,
-            )
-            for a, b in zip(group_A, group_B, strict=True)
-        )
-        prepared = mod._blackwell_grouped_gemm_nt_generated_args(
-            group_A,
-            group_B,
-            out_groups,
-            mma_tiler_mn=(128, 64),
-            cluster_shape_mn=(1, 1),
-        )
-        self.assertIsNone(prepared)
-
-        with patch.object(
-            mod._blackwell_grouped_gemm_nt_generated_kernel,
-            "bind",
-            side_effect=AssertionError("generated bind should not be reached"),
-        ):
-            self.assertIsNone(
-                mod._blackwell_grouped_gemm_nt_generated(
-                    group_A,
-                    group_B,
-                    out_groups,
-                    mma_tiler_mn=(128, 64),
-                    cluster_shape_mn=(1, 1),
-                    use_2cta_instrs=False,
-                )
-            )
-
-        self.assertEqual(grouped_deepgemm._BLACKWELL_GROUPED_COMPILED_CACHE, {})
-        self.assertEqual(grouped_deepgemm._BLACKWELL_GROUPED_PREPARED_CACHE, {})
-
-    def _assert_blackwell_grouped_gemm_nt_correct(
-        self,
-        problem_sizes_mnkl,
-        *,
-        mma_tiler_mn,
-    ) -> None:
-        self._skip_unless_blackwell_grouped_gemm()
-        torch.manual_seed(0)
-        mod = import_path(EXAMPLES_DIR / "grouped_gemm.py")
-        if problem_sizes_mnkl is None:
-            args, expected = mod.make_blackwell_grouped_gemm_nt_args(
-                dtype=torch.float16,
-                device=DEVICE,
-            )
-            problem_sizes_mnkl = mod.BLACKWELL_GROUPED_GEMM_DEFAULT_PROBLEM_SIZES
-        else:
-            args, expected = mod.make_blackwell_grouped_gemm_nt_args(
-                problem_sizes_mnkl,
-                dtype=torch.float16,
-                device=DEVICE,
-            )
-
-        group_A, group_B = args
-        self.assertEqual(
-            tuple(
-                (int(a.size(0)), int(b.size(0)), int(a.size(1)), 1)
-                for a, b in zip(group_A, group_B, strict=True)
-            ),
-            tuple(problem_sizes_mnkl),
-        )
-        actual = mod.blackwell_grouped_gemm_nt(
-            *args,
-            mma_tiler_mn=mma_tiler_mn,
-            cluster_shape_mn=(1, 1),
-        )
-        torch.cuda.synchronize()
-        for group, (actual_group, expected_group) in enumerate(
-            zip(actual, expected, strict=True)
-        ):
-            with self.subTest(group=group):
-                torch.testing.assert_close(
-                    actual_group,
-                    expected_group,
-                    rtol=1e-2,
-                    atol=1e-1,
-                )
-
-    @onlyBackends(["cute"])
-    @skipIfNotCUDA()
-    @patch.dict(
-        os.environ,
-        {
-            "HELION_BACKEND": "cute",
-            "HELION_CUTE_MMA_IMPL": "tcgen05",
-            "HELION_AUTOTUNE_EFFORT": "none",
-        },
-        clear=False,
-    )
-    def test_grouped_gemm_blackwell_nt_official_default_fp16_correct(self):
-        self._assert_blackwell_grouped_gemm_nt_correct(
-            None,
-            mma_tiler_mn=(128, 128),
-        )
-
-    @onlyBackends(["cute"])
-    @skipIfNotCUDA()
-    @patch.dict(
-        os.environ,
-        {
-            "HELION_BACKEND": "cute",
-            "HELION_CUTE_MMA_IMPL": "tcgen05",
-            "HELION_AUTOTUNE_EFFORT": "none",
-        },
-        clear=False,
-    )
-    def test_grouped_gemm_blackwell_nt_documented_mixed_fp16_correct(self):
-        mod = import_path(EXAMPLES_DIR / "grouped_gemm.py")
-        self._assert_blackwell_grouped_gemm_nt_correct(
-            mod.BLACKWELL_GROUPED_GEMM_DOC_PROBLEM_SIZES,
-            mma_tiler_mn=(128, 64),
-        )
-
-    @onlyBackends(["cute"])
-    @skipIfNotCUDA()
-    @patch.dict(
-        os.environ,
-        {
-            "HELION_BACKEND": "cute",
-            "HELION_CUTE_MMA_IMPL": "tcgen05",
-            "HELION_AUTOTUNE_EFFORT": "none",
-        },
-        clear=False,
-    )
-    def test_grouped_gemm_blackwell_nt_generated_explicit_outputs_fp16_correct(self):
-        self._skip_unless_blackwell_grouped_gemm()
-        from helion._compiler.cute import grouped_deepgemm
-
-        grouped_deepgemm._BLACKWELL_GROUPED_COMPILED_CACHE.clear()
-        torch.manual_seed(0)
-        mod = import_path(EXAMPLES_DIR / "grouped_gemm.py")
-        args, expected = mod.make_blackwell_grouped_gemm_nt_args(
-            mod.BLACKWELL_GROUPED_GEMM_DEFAULT_PROBLEM_SIZES,
-            dtype=torch.float16,
-            device=DEVICE,
-        )
-        group_A, group_B = args
-        out_groups = tuple(
-            torch.empty(
-                int(a.size(0)),
-                int(b.size(0)),
-                device=DEVICE,
-                dtype=torch.float16,
-            )
-            for a, b in zip(group_A, group_B, strict=True)
-        )
-
-        actual = mod.blackwell_grouped_gemm_nt_direct(
-            group_A,
-            group_B,
-            out_groups=out_groups,
-            mma_tiler_mn=(128, 64),
-            cluster_shape_mn=(1, 1),
-        )
-        torch.cuda.synchronize()
-        self.assertEqual(grouped_deepgemm._BLACKWELL_GROUPED_COMPILED_CACHE, {})
-        for actual_group, expected_group, out_group in zip(
-            actual,
-            expected,
-            out_groups,
-            strict=True,
-        ):
-            self.assertIs(actual_group, out_group)
-            torch.testing.assert_close(
-                actual_group,
-                expected_group,
-                rtol=1e-2,
-                atol=1e-1,
-            )
-
-    @onlyBackends(["cute"])
-    @skipIfNotCUDA()
-    @patch.dict(
-        os.environ,
-        {
-            "HELION_BACKEND": "cute",
-            "HELION_CUTE_MMA_IMPL": "tcgen05",
-            "HELION_AUTOTUNE_EFFORT": "none",
-        },
-        clear=False,
-    )
-    def test_grouped_gemm_blackwell_nt_generated_explicit_outputs_cache_freshness(
-        self,
-    ):
-        self._skip_unless_blackwell_grouped_gemm()
-
-        torch.manual_seed(0)
-        mod = import_path(EXAMPLES_DIR / "grouped_gemm.py")
-        mod._BLACKWELL_GENERATED_STABLE_LAUNCH_CACHE.clear()
-        args, expected = mod.make_blackwell_grouped_gemm_nt_args(
-            mod.BLACKWELL_GROUPED_GEMM_DEFAULT_PROBLEM_SIZES,
-            dtype=torch.float16,
-            device=DEVICE,
-        )
-        group_A, group_B = args
-        out_groups = tuple(
-            torch.empty(
-                int(a.size(0)),
-                int(b.size(0)),
-                device=DEVICE,
-                dtype=torch.float16,
-            )
-            for a, b in zip(group_A, group_B, strict=True)
-        )
-
-        actual = mod.blackwell_grouped_gemm_nt_direct(
-            group_A,
-            group_B,
-            out_groups=out_groups,
-            mma_tiler_mn=(128, 64),
-            cluster_shape_mn=(1, 1),
-        )
-        torch.cuda.synchronize()
-        self.assertEqual(len(mod._BLACKWELL_GENERATED_STABLE_LAUNCH_CACHE), 1)
-        cached_launch = next(
-            iter(mod._BLACKWELL_GENERATED_STABLE_LAUNCH_CACHE.values())
-        )
-        for actual_group, expected_group in zip(actual, expected, strict=True):
-            torch.testing.assert_close(
-                actual_group,
-                expected_group,
-                rtol=1e-2,
-                atol=1e-1,
-            )
-
-        group_A[0].zero_()
-        updated_expected = mod._reference_blackwell_grouped_gemm_nt(
-            group_A,
-            group_B,
-            out_dtype=torch.float16,
-        )
-        actual = mod.blackwell_grouped_gemm_nt_direct(
-            group_A,
-            group_B,
-            out_groups=out_groups,
-            mma_tiler_mn=(128, 64),
-            cluster_shape_mn=(1, 1),
-        )
-        torch.cuda.synchronize()
-        self.assertIs(
-            next(iter(mod._BLACKWELL_GENERATED_STABLE_LAUNCH_CACHE.values())),
-            cached_launch,
-        )
-        for actual_group, expected_group in zip(
-            actual,
-            updated_expected,
-            strict=True,
-        ):
-            torch.testing.assert_close(
-                actual_group,
-                expected_group,
-                rtol=1e-2,
-                atol=1e-1,
-            )
-
-        fresh_group_A = (group_A[0].clone(), *group_A[1:])
-        actual = mod.blackwell_grouped_gemm_nt_direct(
-            fresh_group_A,
-            group_B,
-            out_groups=out_groups,
-            mma_tiler_mn=(128, 64),
-            cluster_shape_mn=(1, 1),
-        )
-        torch.cuda.synchronize()
-        self.assertEqual(len(mod._BLACKWELL_GENERATED_STABLE_LAUNCH_CACHE), 2)
-        for actual_group, expected_group in zip(
-            actual,
-            updated_expected,
-            strict=True,
-        ):
-            torch.testing.assert_close(
-                actual_group,
-                expected_group,
-                rtol=1e-2,
-                atol=1e-1,
-            )
-
-        fresh_out_groups = tuple(torch.empty_like(out) for out in out_groups)
-        actual = mod.blackwell_grouped_gemm_nt_direct(
-            group_A,
-            group_B,
-            out_groups=fresh_out_groups,
-            mma_tiler_mn=(128, 64),
-            cluster_shape_mn=(1, 1),
-        )
-        torch.cuda.synchronize()
-        self.assertEqual(len(mod._BLACKWELL_GENERATED_STABLE_LAUNCH_CACHE), 3)
-        for actual_group, expected_group, fresh_out in zip(
-            actual,
-            updated_expected,
-            fresh_out_groups,
-            strict=True,
-        ):
-            self.assertIs(actual_group, fresh_out)
-            torch.testing.assert_close(
-                actual_group,
-                expected_group,
-                rtol=1e-2,
-                atol=1e-1,
-            )
-
-        mod._BLACKWELL_GENERATED_STABLE_LAUNCH_CACHE.clear()
-        actual = mod.blackwell_grouped_gemm_nt_direct(
-            group_A,
-            group_B,
-            out_groups=out_groups,
-            mma_tiler_mn=(128, 64),
-            cluster_shape_mn=(1, 1),
-        )
-        torch.cuda.synchronize()
-        self.assertEqual(len(mod._BLACKWELL_GENERATED_STABLE_LAUNCH_CACHE), 1)
-        for actual_group, expected_group in zip(
-            actual,
-            updated_expected,
-            strict=True,
-        ):
-            torch.testing.assert_close(
-                actual_group,
-                expected_group,
-                rtol=1e-2,
-                atol=1e-1,
-            )
-
-    @onlyBackends(["cute"])
-    @skipIfNotCUDA()
-    @patch.dict(
-        os.environ,
-        {
-            "HELION_BACKEND": "cute",
-            "HELION_CUTE_MMA_IMPL": "tcgen05",
-            "HELION_AUTOTUNE_EFFORT": "none",
-        },
-        clear=False,
-    )
-    def test_grouped_gemm_blackwell_nt_generated_explicit_outputs_graph_capture(
-        self,
-    ):
-        self._skip_unless_blackwell_grouped_gemm()
-        from helion._compiler.cute import grouped_deepgemm
-
-        torch.manual_seed(0)
-        mod = import_path(EXAMPLES_DIR / "grouped_gemm.py")
-        mod._BLACKWELL_GENERATED_STABLE_LAUNCH_CACHE.clear()
-        args, expected = mod.make_blackwell_grouped_gemm_nt_args(
-            mod.BLACKWELL_GROUPED_GEMM_DEFAULT_PROBLEM_SIZES,
-            dtype=torch.float16,
-            device=DEVICE,
-        )
-        group_A, group_B = args
-        out_groups = tuple(
-            torch.empty(
-                int(a.size(0)),
-                int(b.size(0)),
-                device=DEVICE,
-                dtype=torch.float16,
-            )
-            for a, b in zip(group_A, group_B, strict=True)
-        )
-
-        mod.blackwell_grouped_gemm_nt_direct(
-            group_A,
-            group_B,
-            out_groups=out_groups,
-            mma_tiler_mn=(128, 64),
-            cluster_shape_mn=(1, 1),
-        )
-        torch.cuda.synchronize()
-        self.assertEqual(len(mod._BLACKWELL_GENERATED_STABLE_LAUNCH_CACHE), 1)
-
-        for out in out_groups:
-            out.fill_(float("nan"))
-        torch.cuda.synchronize()
-        graph = torch.cuda.CUDAGraph()
-        captured: list[tuple[torch.Tensor, ...]] = []
-        with (
-            patch.object(
-                grouped_deepgemm,
-                "blackwell_grouped_gemm_nt",
-                side_effect=AssertionError(
-                    "generated graph capture should not fallback"
-                ),
-            ),
-            torch.cuda.graph(graph),
-        ):
-            captured.append(
-                mod.blackwell_grouped_gemm_nt_direct(
-                    group_A,
-                    group_B,
-                    out_groups=out_groups,
-                    mma_tiler_mn=(128, 64),
-                    cluster_shape_mn=(1, 1),
-                )
-            )
-        for out in out_groups:
-            out.fill_(float("nan"))
-        torch.cuda.synchronize()
-        graph.replay()
-        torch.cuda.synchronize()
-        self.assertEqual(len(captured), 1)
-        for actual_group, expected_group, out_group in zip(
-            captured[0],
-            expected,
-            out_groups,
-            strict=True,
-        ):
-            self.assertIs(actual_group, out_group)
-            torch.testing.assert_close(
-                actual_group,
-                expected_group,
-                rtol=1e-2,
-                atol=1e-1,
-            )
-
-    @onlyBackends(["cute"])
-    @skipIfNotCUDA()
-    @patch.dict(
-        os.environ,
-        {
-            "HELION_BACKEND": "cute",
-            "HELION_CUTE_MMA_IMPL": "tcgen05",
-            "HELION_AUTOTUNE_EFFORT": "none",
-        },
-        clear=False,
-    )
-    def test_blackwell_nt_generated_explicit_outputs_graph_capture_lru_hit(
-        self,
-    ):
-        self._skip_unless_blackwell_grouped_gemm()
-        from helion._compiler.cute import grouped_deepgemm
-
-        torch.manual_seed(0)
-        mod = import_path(EXAMPLES_DIR / "grouped_gemm.py")
-        mod._BLACKWELL_GENERATED_STABLE_LAUNCH_CACHE.clear()
-        args, _expected = mod.make_blackwell_grouped_gemm_nt_args(
-            mod.BLACKWELL_GROUPED_GEMM_DEFAULT_PROBLEM_SIZES,
-            dtype=torch.float16,
-            device=DEVICE,
-        )
-        group_A, group_B = args
-
-        def new_out_groups() -> tuple[torch.Tensor, ...]:
-            return tuple(
-                torch.empty(
-                    int(a.size(0)),
-                    int(b.size(0)),
-                    device=DEVICE,
-                    dtype=torch.float16,
-                )
-                for a, b in zip(group_A, group_B, strict=True)
-            )
-
-        out_first = new_out_groups()
-        out_second = new_out_groups()
-        mod.blackwell_grouped_gemm_nt_direct(
-            group_A,
-            group_B,
-            out_groups=out_first,
-            mma_tiler_mn=(128, 64),
-            cluster_shape_mn=(1, 1),
-        )
-        torch.cuda.synchronize()
-        self.assertEqual(len(mod._BLACKWELL_GENERATED_STABLE_LAUNCH_CACHE), 1)
-        first_key = next(iter(mod._BLACKWELL_GENERATED_STABLE_LAUNCH_CACHE))
-        first_launch = mod._BLACKWELL_GENERATED_STABLE_LAUNCH_CACHE[first_key]
-
-        mod.blackwell_grouped_gemm_nt_direct(
-            group_A,
-            group_B,
-            out_groups=out_second,
-            mma_tiler_mn=(128, 64),
-            cluster_shape_mn=(1, 1),
-        )
-        torch.cuda.synchronize()
-        self.assertEqual(len(mod._BLACKWELL_GENERATED_STABLE_LAUNCH_CACHE), 2)
-        self.assertIn(first_key, mod._BLACKWELL_GENERATED_STABLE_LAUNCH_CACHE)
-        self.assertIsNot(mod._BLACKWELL_GENERATED_LAST_STABLE_LAUNCH, first_launch)
-
-        group_A[0].zero_()
-        expected = mod._reference_blackwell_grouped_gemm_nt(
-            group_A,
-            group_B,
-            out_dtype=torch.float16,
-        )
-        torch.cuda.synchronize()
-
-        for out in out_first:
-            out.fill_(float("nan"))
-        torch.cuda.synchronize()
-        graph = torch.cuda.CUDAGraph()
-        captured: list[tuple[torch.Tensor, ...]] = []
-        with (
-            patch.object(
-                grouped_deepgemm,
-                "blackwell_grouped_gemm_nt",
-                side_effect=AssertionError(
-                    "generated graph capture should not fallback"
-                ),
-            ),
-            patch.object(
-                mod._blackwell_grouped_gemm_nt_generated_kernel,
-                "bind",
-                side_effect=AssertionError("cached graph capture should not rebind"),
-            ),
-            torch.cuda.graph(graph),
-        ):
-            captured.append(
-                mod.blackwell_grouped_gemm_nt_direct(
-                    group_A,
-                    group_B,
-                    out_groups=out_first,
-                    mma_tiler_mn=(128, 64),
-                    cluster_shape_mn=(1, 1),
-                )
-            )
-
-        for out in out_first:
-            out.fill_(float("nan"))
-        torch.cuda.synchronize()
-        graph.replay()
-        torch.cuda.synchronize()
-
-        self.assertEqual(len(captured), 1)
-        self.assertIs(mod._BLACKWELL_GENERATED_LAST_STABLE_LAUNCH, first_launch)
-        for actual_group, expected_group, out_group in zip(
-            captured[0],
-            expected,
-            out_first,
-            strict=True,
-        ):
-            self.assertIs(actual_group, out_group)
-            torch.testing.assert_close(
-                actual_group,
-                expected_group,
-                rtol=1e-2,
-                atol=1e-1,
-            )
-
-    @onlyBackends(["cute"])
-    @skipIfNotCUDA()
-    @patch.dict(
-        os.environ,
-        {
-            "HELION_BACKEND": "cute",
-            "HELION_CUTE_MMA_IMPL": "tcgen05",
-            "HELION_AUTOTUNE_EFFORT": "none",
-        },
-        clear=False,
-    )
-    def test_grouped_gemm_blackwell_nt_generated_explicit_outputs_retains_owned_metadata(
-        self,
-    ):
-        self._skip_unless_blackwell_grouped_gemm()
-        import helion.runtime as helion_runtime
-
-        torch.manual_seed(0)
-        mod = import_path(EXAMPLES_DIR / "grouped_gemm.py")
-        mod._BLACKWELL_GENERATED_STABLE_LAUNCH_CACHE.clear()
-        args, expected = mod.make_blackwell_grouped_gemm_nt_args(
-            mod.BLACKWELL_GROUPED_GEMM_DEFAULT_PROBLEM_SIZES,
-            dtype=torch.float16,
-            device=DEVICE,
-        )
-        group_A, group_B = args
-
-        def new_out_groups() -> tuple[torch.Tensor, ...]:
-            return tuple(
-                torch.empty(
-                    int(a.size(0)),
-                    int(b.size(0)),
-                    device=DEVICE,
-                    dtype=torch.float16,
-                )
-                for a, b in zip(group_A, group_B, strict=True)
-            )
-
-        out_groups = new_out_groups()
-        actual = mod.blackwell_grouped_gemm_nt_direct(
-            group_A,
-            group_B,
-            out_groups=out_groups,
-            mma_tiler_mn=(128, 64),
-            cluster_shape_mn=(1, 1),
-        )
-        torch.cuda.synchronize()
-        self.assertEqual(len(mod._BLACKWELL_GENERATED_STABLE_LAUNCH_CACHE), 1)
-        first_launch = next(iter(mod._BLACKWELL_GENERATED_STABLE_LAUNCH_CACHE.values()))
-        self.assertIsNotNone(first_launch.fast_call)
-        self.assertIsNotNone(first_launch.runtime_cache_entry)
-        runtime_cache_entry = first_launch.runtime_cache_entry
-        retained_owned_tensors = runtime_cache_entry.owned_tensors
-        self.assertGreater(len(retained_owned_tensors), 0)
-        retained_owned_ptrs = tuple(
-            int(tensor.data_ptr()) for tensor in retained_owned_tensors
-        )
-        for actual_group, expected_group in zip(actual, expected, strict=True):
-            torch.testing.assert_close(
-                actual_group,
-                expected_group,
-                rtol=1e-2,
-                atol=1e-1,
-            )
-
-        pressure_launches = helion_runtime._CUTE_LAUNCH_ARG_CACHE_LIMIT + 2
-        self.assertLess(
-            pressure_launches + 1,
-            mod._BLACKWELL_GENERATED_STABLE_LAUNCH_CACHE_MAX,
-        )
-        pressure_out_groups = []
-        for _ in range(pressure_launches):
-            pressure_out = new_out_groups()
-            pressure_out_groups.append(pressure_out)
-            mod.blackwell_grouped_gemm_nt_direct(
-                group_A,
-                group_B,
-                out_groups=pressure_out,
-                mma_tiler_mn=(128, 64),
-                cluster_shape_mn=(1, 1),
-            )
-        torch.cuda.synchronize()
-        self.assertIn(
-            first_launch.cache_key, mod._BLACKWELL_GENERATED_STABLE_LAUNCH_CACHE
-        )
-        self.assertIs(first_launch.runtime_cache_entry, runtime_cache_entry)
-        self.assertEqual(
-            tuple(int(tensor.data_ptr()) for tensor in retained_owned_tensors),
-            retained_owned_ptrs,
-        )
-
-        junk_tensors = []
-        for fill_value in range(64):
-            for tensor in retained_owned_tensors:
-                junk = torch.empty_like(tensor)
-                junk.fill_(fill_value + 1024)
-                junk_tensors.append(junk)
-
-        group_A[0].zero_()
-        updated_expected = mod._reference_blackwell_grouped_gemm_nt(
-            group_A,
-            group_B,
-            out_dtype=torch.float16,
-        )
-        for out in out_groups:
-            out.fill_(float("nan"))
-        torch.cuda.synchronize()
-        with patch.object(
-            mod._blackwell_grouped_gemm_nt_generated_kernel,
-            "bind",
-            side_effect=AssertionError("relaunch should use cached public launch"),
-        ):
-            actual = mod.blackwell_grouped_gemm_nt_direct(
-                group_A,
-                group_B,
-                out_groups=out_groups,
-                mma_tiler_mn=(128, 64),
-                cluster_shape_mn=(1, 1),
-            )
-        torch.cuda.synchronize()
-        for actual_group, expected_group, out_group in zip(
-            actual,
-            updated_expected,
-            out_groups,
-            strict=True,
-        ):
-            self.assertIs(actual_group, out_group)
-            torch.testing.assert_close(
-                actual_group,
-                expected_group,
-                rtol=1e-2,
-                atol=1e-1,
-            )
-        self.assertEqual(len(junk_tensors), 64 * len(retained_owned_tensors))
-        self.assertEqual(len(pressure_out_groups), pressure_launches)
-
-    @onlyBackends(["cute"])
-    @skipIfNotCUDA()
-    @patch.dict(
-        os.environ,
-        {
-            "HELION_BACKEND": "cute",
-            "HELION_CUTE_MMA_IMPL": "tcgen05",
-            "HELION_AUTOTUNE_EFFORT": "none",
-        },
-        clear=False,
-    )
-    def test_grouped_gemm_blackwell_nt_generated_documented_fp16_bf16_correct(self):
-        self._skip_unless_blackwell_grouped_gemm()
-        from helion._compiler.cute import grouped_deepgemm
-
-        mod = import_path(EXAMPLES_DIR / "grouped_gemm.py")
-        for dtype, rtol, atol in (
-            (torch.float16, 4e-2, 1e-1),
-            (torch.bfloat16, 4e-2, 2e-1),
-        ):
-            with self.subTest(dtype=dtype):
-                grouped_deepgemm._BLACKWELL_GROUPED_COMPILED_CACHE.clear()
-                torch.manual_seed(0)
-                args, expected = mod.make_blackwell_grouped_gemm_nt_args(
-                    mod.BLACKWELL_GROUPED_GEMM_DOC_PROBLEM_SIZES,
-                    dtype=dtype,
-                    device=DEVICE,
-                )
-                actual = mod.blackwell_grouped_gemm_nt_direct(
-                    *args,
-                    mma_tiler_mn=(128, 64),
-                    cluster_shape_mn=(1, 1),
-                )
-                torch.cuda.synchronize()
-                self.assertEqual(
-                    grouped_deepgemm._BLACKWELL_GROUPED_COMPILED_CACHE,
-                    {},
-                )
-                for actual_group, expected_group in zip(
-                    actual,
-                    expected,
-                    strict=True,
-                ):
-                    torch.testing.assert_close(
-                        actual_group,
-                        expected_group,
-                        rtol=rtol,
-                        atol=atol,
-                    )
-
-    @onlyBackends(["cute"])
-    @skipIfNotCUDA()
-    @patch.dict(
-        os.environ,
-        {
-            "HELION_BACKEND": "cute",
-            "HELION_CUTE_MMA_IMPL": "tcgen05",
-            "HELION_AUTOTUNE_EFFORT": "none",
-        },
-        clear=False,
-    )
-    def test_grouped_gemm_blackwell_nt_prepares_stable_output_launch(self):
-        self._skip_unless_blackwell_grouped_gemm()
-        from helion._compiler.cute import grouped_deepgemm
-
-        grouped_deepgemm._BLACKWELL_GROUPED_PREPARED_CACHE.clear()
-        torch.manual_seed(0)
-        mod = import_path(EXAMPLES_DIR / "grouped_gemm.py")
-        args, expected = mod.make_blackwell_grouped_gemm_nt_args(
-            ((128, 128, 128, 1),),
-            dtype=torch.float16,
-            device=DEVICE,
-        )
-        group_A, group_B = args
-        out_groups = tuple(
-            torch.empty(
-                int(a.size(0)),
-                int(b.size(0)),
-                device=DEVICE,
-                dtype=torch.float16,
-            )
-            for a, b in zip(group_A, group_B, strict=True)
-        )
-
-        actual = mod.blackwell_grouped_gemm_nt(
-            group_A,
-            group_B,
-            out_groups=out_groups,
-            mma_tiler_mn=(128, 128),
-            cluster_shape_mn=(1, 1),
-        )
-        torch.cuda.synchronize()
-        torch.testing.assert_close(actual[0], expected[0], rtol=1e-2, atol=1e-1)
-        self.assertEqual(len(grouped_deepgemm._BLACKWELL_GROUPED_PREPARED_CACHE), 1)
-        prepared = next(
-            iter(grouped_deepgemm._BLACKWELL_GROUPED_PREPARED_CACHE.values())
-        )
-        self.assertEqual(len(prepared.stream_executions), 1)
-
-        actual = mod.blackwell_grouped_gemm_nt(
-            group_A,
-            group_B,
-            out_groups=out_groups,
-            mma_tiler_mn=(128, 128),
-            cluster_shape_mn=(1, 1),
-        )
-        torch.cuda.synchronize()
-        torch.testing.assert_close(actual[0], expected[0], rtol=1e-2, atol=1e-1)
-        self.assertIs(
-            next(iter(grouped_deepgemm._BLACKWELL_GROUPED_PREPARED_CACHE.values())),
-            prepared,
-        )
-
-        graph = torch.cuda.CUDAGraph()
-        torch.cuda.synchronize()
-        with torch.cuda.graph(graph):
-            mod.blackwell_grouped_gemm_nt(
-                group_A,
-                group_B,
-                out_groups=out_groups,
-                mma_tiler_mn=(128, 128),
-                cluster_shape_mn=(1, 1),
-            )
-        graph.replay()
-        torch.cuda.synchronize()
-        torch.testing.assert_close(out_groups[0], expected[0], rtol=1e-2, atol=1e-1)
-
-    def _skip_unless_deepgemm_segment_tcgen05(self) -> None:
-        from helion._compat import requires_cuda_version
-        from helion._compiler.cute.mma_support import get_cute_mma_support
-
-        if os.environ.get("HELION_CUTE_MMA_IMPL", "").strip().lower() != "tcgen05":
-            self.skipTest("CuTeDSL generated segment path requires tcgen05")
-        if not requires_cuda_version("13"):
-            self.skipTest("CuTeDSL generated segment path requires CUDA >= 13")
-        with torch.cuda.device(DEVICE):
-            major, _minor = torch.cuda.get_device_capability(DEVICE)
-            if major < 10:
-                self.skipTest("CuTeDSL generated segment path requires SM100+")
-            if not get_cute_mma_support().tcgen05_f16bf16:
-                self.skipTest("tcgen05 F16/BF16 MMA is not supported on this machine")
-
-    def _deepgemm_segment_metadata_fixture(
-        self,
-    ) -> tuple[tuple[torch.Tensor, torch.Tensor, torch.Tensor], torch.Tensor]:
-        torch.manual_seed(0)
-        mod = import_path(EXAMPLES_DIR / "grouped_gemm.py")
-        args, _expected = mod.make_deepgemm_m_grouped_bf16_gemm_nt_contiguous_args(
-            (188, 64, 40, 64),
-            n=128,
-            k=64,
-            m_alignment=224,
-            tail_padding=0,
-            device=DEVICE,
-        )
-        work_tile_metadata = (
-            mod._deepgemm_m_grouped_bf16_gemm_nt_contiguous_segment_metadata_tensor(
-                *args
-            )
-        )
-        self.assertIsNotNone(work_tile_metadata)
-        return args, work_tile_metadata
-
-    def _patch_torch_empty_for_deepgemm_output(
-        self,
-        replacement: torch.Tensor,
-    ) -> object:
-        real_empty = torch.empty
-
-        def patched_empty(*shape: object, **kwargs: object) -> torch.Tensor:
-            requested_shape = shape
-            if len(shape) == 1 and isinstance(shape[0], (tuple, list, torch.Size)):
-                requested_shape = tuple(shape[0])
-            requested_device = kwargs.get("device")
-            if (
-                tuple(int(dim) for dim in requested_shape) == tuple(replacement.shape)
-                and kwargs.get("dtype") == replacement.dtype
-                and requested_device is not None
-                and torch.device(requested_device) == replacement.device
-            ):
-                return replacement
-            return real_empty(*shape, **kwargs)
-
-        return patch.object(torch, "empty", side_effect=patched_empty)
-
-    @onlyBackends(["cute"])
-    @skipIfNotCUDA()
-    @skipIfRefEager("selected generated segment requires compiled CuTe code")
-    @patch.dict(
-        os.environ,
-        {
-            "HELION_BACKEND": "cute",
-            "HELION_CUTE_MMA_IMPL": "tcgen05",
-            "HELION_AUTOTUNE_EFFORT": "none",
-        },
-        clear=False,
-    )
-    def test_grouped_gemm_deepgemm_generated_selected_segment_source_and_route(
-        self,
-    ):
-        self._skip_unless_deepgemm_segment_tcgen05()
-        torch.manual_seed(0)
-        mod = import_path(EXAMPLES_DIR / "grouped_gemm.py")
-        args, expected = mod.make_deepgemm_m_grouped_bf16_gemm_nt_contiguous_args(
-            (188, 64, 40, 64),
-            n=128,
-            k=64,
-            m_alignment=224,
-            tail_padding=0,
-            device=DEVICE,
-        )
-        A_packed, B_grouped, grouped_layout = args
-        A_packed[grouped_layout == -1] = 0
-        route_report = mod.deepgemm_m_grouped_bf16_gemm_nt_contiguous_route_report(
-            *args
-        )
-        self.assertEqual(route_report.route, "generated_selected_segment")
-        self.assertEqual(
-            route_report.as_dict(),
-            {
-                "layout_has_valid_prefix_tiles": False,
-                "use_generated_segment_kernel": True,
-                "use_selected_segment_kernel": True,
-                "selected_metadata_nonempty": True,
-                "route": "generated_selected_segment",
-            },
-        )
-        work_tile_metadata = route_report.selected_segment_work_tile_metadata
-        self.assertIsNotNone(work_tile_metadata)
-        assert work_tile_metadata is not None
-        self.assertEqual(
-            work_tile_metadata.detach().cpu().tolist(),
-            [
-                [0, 0, 188, 224],
-                [1, 224, 64, 224],
-                [2, 448, 40, 224],
-                [3, 672, 64, 224],
-            ],
-        )
-
-        selected_kernel = (
-            mod._deepgemm_m_grouped_bf16_gemm_nt_contiguous_selected_segment_kernel
-        )
-        bound = selected_kernel.bind((A_packed, B_grouped, work_tile_metadata))
-        bound.env.config_spec.cute_tcgen05_search_enabled = True
-        (config,) = selected_kernel.configs
-        self.assertEqual(config["block_sizes"], [256, 128, 64])
-        self.assertEqual(config["tcgen05_cluster_m"], 2)
-        self.assertEqual(config["tcgen05_cluster_n"], 1)
-        self.assertEqual(config["tcgen05_ab_stages"], 7)
-        self.assertTrue(config["tcgen05_deepgemm_selected"])
-        self.assertTrue(config["tcgen05_deepgemm_selected_compact_metadata"])
-        self.assertEqual(config["tcgen05_selected_accumulator_view"], "nm")
-        self.assertEqual(config["tcgen05_selected_d_store_view"], "nm_transposed")
-        code = bound.to_triton_code(config)
-        self.assertIn("tcgen05_grouped_selected_sched", code)
-        self.assertIn("tcgen05_bSG_gD[None, cutlass.Int32(_tcgen05_subtile)]", code)
-        self.assertIn("cute.nvgpu.tcgen05.CtaGroup.TWO", code)
-        self.assertNotIn("CtaGroup.ONE", code)
-        self.assertIn("out = torch.empty(", code)
-        self.assertNotIn("out = torch.zeros(", code)
-        self.assertNotIn(".zero_(", code)
-        for helper_marker in (
-            "_cutlass_grouped_gemm_kernel",
-            "GroupedGemmKernel",
-            "grouped_deepgemm",
-            "_deepgemm_m_grouped_bf16_gemm_nt_contiguous_zero_padding_rows",
-            "_deepgemm_m_grouped_bf16_gemm_nt_contiguous_selected_segment_output",
-        ):
-            self.assertNotIn(helper_marker, code)
-        self.assertFalse(
-            hasattr(
-                mod,
-                "_deepgemm_m_grouped_bf16_gemm_nt_contiguous_zero_padding_rows",
-            )
-        )
-        self.assertFalse(
-            hasattr(
-                mod,
-                "_deepgemm_m_grouped_bf16_gemm_nt_contiguous_selected_segment_output",
-            )
-        )
-
-        grouped_plan = next(
-            plan
-            for plan in _cute_wrapper_plans_from_code(code)
-            if plan["kind"] == "tcgen05_grouped_static_persistent"
-        )
-        self.assertIs(grouped_plan["deepgemm_selected"], True)
-        self.assertIs(grouped_plan["deepgemm_selected_compact_metadata"], True)
-        self.assertEqual(grouped_plan["source_m_tile"], 224)
-        self.assertEqual(grouped_plan["cluster_m"], 2)
-        self.assertEqual(grouped_plan["cluster_n"], 1)
-        self.assertEqual(grouped_plan["selected_store_wave"], "nm_explicit_128x32")
-        ab_plan = next(
-            plan
-            for plan in _cute_wrapper_plans_from_code(code)
-            if plan["kind"] == "tcgen05_ab_tma"
-        )
-        self.assertEqual(ab_plan["ab_stage_count"], 7)
-        selected_kernel.reset()
-
-        with patch.object(
-            mod,
-            "_deepgemm_m_grouped_bf16_gemm_nt_contiguous_segment_metadata_tensor",
-            side_effect=AssertionError("old segment fallback must not run"),
-        ):
-            actual = mod.deepgemm_m_grouped_bf16_gemm_nt_contiguous(*args)
-            torch.cuda.synchronize()
-        torch.testing.assert_close(actual, expected, rtol=2e-2, atol=2e-2)
-        self.assertTrue(bool(torch.all(actual[grouped_layout == -1] == 0).item()))
-
-        sentinel = torch.full_like(expected, -77.0)
-        with (
-            patch.object(
-                mod,
-                "_deepgemm_m_grouped_bf16_gemm_nt_contiguous_segment_metadata_tensor",
-                side_effect=AssertionError("old segment fallback must not run"),
-            ),
-            self._patch_torch_empty_for_deepgemm_output(sentinel),
-        ):
-            actual = mod.deepgemm_m_grouped_bf16_gemm_nt_contiguous(*args)
-            torch.cuda.synchronize()
-        self.assertIs(actual, sentinel)
-        torch.testing.assert_close(
-            actual[grouped_layout != -1],
-            expected[grouped_layout != -1],
-            rtol=2e-2,
-            atol=2e-2,
-        )
-        self.assertTrue(bool(torch.all(actual[grouped_layout == -1] == 0).item()))
-
-    @skipIfRefEager("Route report metadata inspection does not execute a Helion kernel")
-    def test_grouped_gemm_deepgemm_route_report_metadata_does_not_alias_cache(self):
-        torch.manual_seed(0)
-        mod = import_path(EXAMPLES_DIR / "grouped_gemm.py")
-        args, _expected = mod.make_deepgemm_m_grouped_bf16_gemm_nt_contiguous_args(
-            (188, 64, 40, 64),
-            n=128,
-            k=64,
-            m_alignment=224,
-            tail_padding=0,
-            device=torch.device("cpu"),
-        )
-        route_report = mod.deepgemm_m_grouped_bf16_gemm_nt_contiguous_route_report(
-            *args
-        )
-        reported_metadata = route_report.selected_segment_work_tile_metadata
-        self.assertIsNotNone(reported_metadata)
-        assert reported_metadata is not None
-
-        cached_metadata = mod._deepgemm_m_grouped_bf16_gemm_nt_contiguous_selected_segment_metadata_tensor(
-            *args
-        )
-        self.assertIsNotNone(cached_metadata)
-        assert cached_metadata is not None
-        self.assertIsNot(reported_metadata, cached_metadata)
-        expected_metadata = cached_metadata.detach().cpu().tolist()
-
-        reported_metadata[0, 0] = 99
-        self.assertEqual(cached_metadata.detach().cpu().tolist(), expected_metadata)
-
-        second_report = mod.deepgemm_m_grouped_bf16_gemm_nt_contiguous_route_report(
-            *args
-        )
-        second_metadata = second_report.selected_segment_work_tile_metadata
-        self.assertIsNotNone(second_metadata)
-        assert second_metadata is not None
-        self.assertEqual(second_metadata.detach().cpu().tolist(), expected_metadata)
-
-    @onlyBackends(["cute"])
-    @skipIfNotCUDA()
-    @skipIfRefEager("CUDA graph capture requires compiled kernel")
-    @patch.dict(
-        os.environ,
-        {
-            "HELION_BACKEND": "cute",
-            "HELION_CUTE_MMA_IMPL": "tcgen05",
-            "HELION_AUTOTUNE_EFFORT": "none",
-        },
-        clear=False,
-    )
-    def test_grouped_gemm_deepgemm_generated_selected_segment_graph_capture_padding(
-        self,
-    ):
-        self._skip_unless_deepgemm_segment_tcgen05()
-        torch.manual_seed(0)
-        mod = import_path(EXAMPLES_DIR / "grouped_gemm.py")
-        args, expected = mod.make_deepgemm_m_grouped_bf16_gemm_nt_contiguous_args(
-            (188, 64, 40, 64),
-            n=128,
-            k=64,
-            m_alignment=224,
-            tail_padding=0,
-            device=DEVICE,
-        )
-        A_packed, _B_grouped, grouped_layout = args
-        padding_rows = grouped_layout == -1
-        valid_rows = ~padding_rows
-        self.assertGreater(int(torch.count_nonzero(padding_rows).item()), 0)
-        self.assertTrue(
-            mod._deepgemm_m_grouped_bf16_gemm_nt_contiguous_use_selected_segment_kernel(
-                *args
-            )
-        )
-        with patch.object(
-            mod,
-            "_deepgemm_m_grouped_bf16_gemm_nt_contiguous_segment_metadata_tensor",
-            side_effect=AssertionError("old segment fallback must not run"),
-        ):
-            warm = mod.deepgemm_m_grouped_bf16_gemm_nt_contiguous(*args)
-            torch.cuda.synchronize()
-        torch.testing.assert_close(warm, expected, rtol=2e-2, atol=2e-2)
-
-        A_packed[padding_rows] = 7
-        torch.cuda.synchronize()
-
-        graph = torch.cuda.CUDAGraph()
-        torch.cuda.synchronize()
-        with (
-            patch.object(
-                mod,
-                "_deepgemm_m_grouped_bf16_gemm_nt_contiguous_segment_metadata_tensor",
-                side_effect=AssertionError("old segment fallback must not run"),
-            ),
-            torch.cuda.graph(graph),
-        ):
-            captured = mod.deepgemm_m_grouped_bf16_gemm_nt_contiguous(*args)
-        torch.cuda.synchronize()
-
-        captured[valid_rows] = -5
-        captured[padding_rows] = 13
-        torch.cuda.synchronize()
-        graph.replay()
-        torch.cuda.synchronize()
-        torch.testing.assert_close(
-            captured[valid_rows],
-            expected[valid_rows],
-            rtol=2e-2,
-            atol=2e-2,
-        )
-        self.assertTrue(bool(torch.all(captured[padding_rows] == 0).item()))
-
-    @onlyBackends(["cute"])
-    @skipIfNotCUDA()
-    @skipIfRefEager("selected generated segment requires compiled CuTe code")
-    @patch.dict(
-        os.environ,
-        {
-            "HELION_BACKEND": "cute",
-            "HELION_CUTE_MMA_IMPL": "tcgen05",
-            "HELION_AUTOTUNE_EFFORT": "none",
-        },
-        clear=False,
-    )
-    def test_grouped_gemm_deepgemm_generated_selected_segment_full_overwrite(
-        self,
-    ):
-        self._skip_unless_deepgemm_segment_tcgen05()
-        torch.manual_seed(0)
-        mod = import_path(EXAMPLES_DIR / "grouped_gemm.py")
-        args, expected = mod.make_deepgemm_m_grouped_bf16_gemm_nt_contiguous_args(
-            (224, 224),
-            n=128,
-            k=64,
-            m_alignment=224,
-            tail_padding=0,
-            device=DEVICE,
-        )
-        _A_packed, _B_grouped, grouped_layout = args
-        self.assertEqual(int(torch.count_nonzero(grouped_layout == -1).item()), 0)
-        self.assertTrue(
-            mod._deepgemm_m_grouped_bf16_gemm_nt_contiguous_use_selected_segment_kernel(
-                *args
-            )
-        )
-        work_tile_metadata = mod._deepgemm_m_grouped_bf16_gemm_nt_contiguous_selected_segment_metadata_tensor(
-            *args
-        )
-        self.assertIsNotNone(work_tile_metadata)
-        assert work_tile_metadata is not None
-        self.assertEqual(
-            work_tile_metadata.detach().cpu().tolist(),
-            [
-                [0, 0, 224, 224],
-                [1, 224, 224, 224],
-            ],
-        )
-
-        with patch.object(
-            mod,
-            "_deepgemm_m_grouped_bf16_gemm_nt_contiguous_segment_metadata_tensor",
-            side_effect=AssertionError("old segment fallback must not run"),
-        ):
-            warm = mod.deepgemm_m_grouped_bf16_gemm_nt_contiguous(*args)
-            torch.cuda.synchronize()
-        torch.testing.assert_close(warm, expected, rtol=2e-2, atol=2e-2)
-
-        sentinel = torch.full_like(expected, float("nan"))
-        with (
-            patch.object(
-                mod,
-                "_deepgemm_m_grouped_bf16_gemm_nt_contiguous_segment_metadata_tensor",
-                side_effect=AssertionError("old segment fallback must not run"),
-            ),
-            self._patch_torch_empty_for_deepgemm_output(sentinel),
-        ):
-            actual = mod.deepgemm_m_grouped_bf16_gemm_nt_contiguous(*args)
-            torch.cuda.synchronize()
-        self.assertIs(actual, sentinel)
-        self.assertFalse(bool(torch.any(torch.isnan(actual)).item()))
-        torch.testing.assert_close(actual, expected, rtol=2e-2, atol=2e-2)
-
-    @onlyBackends(["cute"])
-    @skipIfNotCUDA()
-    @skipIfRefEager("selected generated segment requires compiled CuTe code")
-    @patch.dict(
-        os.environ,
-        {
-            "HELION_BACKEND": "cute",
-            "HELION_CUTE_MMA_IMPL": "tcgen05",
-            "HELION_AUTOTUNE_EFFORT": "none",
-        },
-        clear=False,
-    )
-    def test_grouped_gemm_deepgemm_generated_selected_segment_padding_only_chunk(
-        self,
-    ):
-        self._skip_unless_deepgemm_segment_tcgen05()
-        torch.manual_seed(0)
-        mod = import_path(EXAMPLES_DIR / "grouped_gemm.py")
-        args, expected = mod.make_deepgemm_m_grouped_bf16_gemm_nt_contiguous_args(
-            (224,),
-            n=128,
-            k=64,
-            m_alignment=448,
-            tail_padding=0,
-            device=DEVICE,
-        )
-        A_packed, _B_grouped, grouped_layout = args
-        A_packed[grouped_layout == -1] = 7
-        self.assertTrue(
-            mod._deepgemm_m_grouped_bf16_gemm_nt_contiguous_use_selected_segment_kernel(
-                *args
-            )
-        )
-        work_tile_metadata = mod._deepgemm_m_grouped_bf16_gemm_nt_contiguous_selected_segment_metadata_tensor(
-            *args
-        )
-        self.assertIsNotNone(work_tile_metadata)
-        assert work_tile_metadata is not None
-        self.assertEqual(
-            work_tile_metadata.detach().cpu().tolist(),
-            [
-                [0, 0, 224, 448],
-            ],
-        )
-
-        with patch.object(
-            mod,
-            "_deepgemm_m_grouped_bf16_gemm_nt_contiguous_segment_metadata_tensor",
-            side_effect=AssertionError("old segment fallback must not run"),
-        ):
-            actual = mod.deepgemm_m_grouped_bf16_gemm_nt_contiguous(*args)
-            torch.cuda.synchronize()
-        torch.testing.assert_close(
-            actual[grouped_layout != -1],
-            expected[grouped_layout != -1],
-            rtol=2e-2,
-            atol=2e-2,
-        )
-        self.assertTrue(bool(torch.all(actual[grouped_layout == -1] == 0).item()))
-
-    def _assert_segment_tcgen05_not_admitted(
-        self,
-        kernel: helion.Kernel,
-        args: tuple[torch.Tensor, torch.Tensor, torch.Tensor],
-        work_tile_metadata: torch.Tensor,
-    ) -> None:
-        bound = kernel.bind((args[0], args[1], work_tile_metadata))
-        bound.env.config_spec.cute_tcgen05_search_enabled = True
-        (config,) = kernel.configs
-        code = bound.to_triton_code(config)
-        self.assertNotIn("cute.nvgpu.tcgen05", code)
-        self.assertNotIn("cute.gemm(", code)
-
-    @onlyBackends(["cute"])
-    @skipIfNotCUDA()
-    @skipIfRefEager("CUDA graph capture requires compiled kernel")
-    @patch.dict(
-        os.environ,
-        {
-            "HELION_BACKEND": "cute",
-            "HELION_CUTE_MMA_IMPL": "tcgen05",
-            "HELION_AUTOTUNE_EFFORT": "none",
-        },
-        clear=False,
-    )
-    def test_grouped_gemm_deepgemm_generated_segment_mixed_boundary(
-        self,
-    ):
-        self._skip_unless_deepgemm_segment_tcgen05()
-        torch.manual_seed(0)
-        mod = import_path(EXAMPLES_DIR / "grouped_gemm.py")
-        args, expected = mod.make_deepgemm_m_grouped_bf16_gemm_nt_contiguous_args(
-            (188, 64, 40, 64),
-            n=128,
-            k=64,
-            m_alignment=224,
-            tail_padding=0,
-            device=DEVICE,
-        )
-        layout_has_valid_prefix_tiles = (
-            mod._validate_deepgemm_m_grouped_bf16_gemm_nt_contiguous_args(*args)
-        )
-        self.assertFalse(layout_has_valid_prefix_tiles)
-        self.assertTrue(
-            mod._deepgemm_m_grouped_bf16_gemm_nt_contiguous_use_generated_segment_kernel(
-                *args,
-                layout_has_valid_prefix_tiles=layout_has_valid_prefix_tiles,
-            )
-        )
-        self.assertFalse(
-            mod._deepgemm_m_grouped_bf16_gemm_nt_contiguous_segment_output_fully_overwritten(
-                *args
-            )
-        )
-
-        work_tile_metadata = (
-            mod._deepgemm_m_grouped_bf16_gemm_nt_contiguous_segment_metadata_tensor(
-                *args
-            )
-        )
-        self.assertIsNotNone(work_tile_metadata)
-        self.assertEqual(
-            work_tile_metadata.detach().cpu().tolist(),
-            [
-                [0, 0, 128, 0],
-                [0, 128, 60, 0],
-                [1, 224, 64, 0],
-                [2, 448, 40, 0],
-                [3, 672, 64, 0],
-            ],
-        )
-        segment_kernel = mod._deepgemm_m_grouped_bf16_gemm_nt_contiguous_segment_kernel
-        bound = segment_kernel.bind((args[0], args[1], work_tile_metadata))
-        bound.env.config_spec.cute_tcgen05_search_enabled = True
-        (config,) = segment_kernel.configs
-        self.assertEqual(config["block_sizes"], [128, 128, 64])
-        self.assertEqual(config["tcgen05_ab_stages"], 3)
-        self.assertEqual(config["tcgen05_c_stages"], 4)
-        code = bound.to_triton_code(config)
-        self.assertIn("'cta_tile_shape_mnk': (128, 128, 64)", code)
-        self.assertIn("'bn': 128", code)
-        self.assertIn("'ab_stage_count': 3", code)
-        self.assertIn("'c_stage_count': 4", code)
-        self.assertIn("cute.nvgpu.tcgen05", code)
-        self.assertIn("cute.gemm(", code)
-        self.assertIn("PipelineTmaUmma.create", code)
-        self.assertIn("StaticPersistentGroupTileScheduler.create", code)
-        self.assertIn("TensorMapManager", code)
-        self.assertIn("update_tensormap", code)
-        self.assertIn("tma_partition", code)
-        self.assertIn("CtaGroup.ONE", code)
-        self.assertNotIn("CtaGroup.TWO", code)
-        self.assertNotIn("_cute_grouped_reduce_shared_two_stage", code)
-        self.assertNotIn("_cutlass_grouped_gemm_kernel", code)
-        self.assertNotIn("dot_product", code)
-        self.assertNotIn("for _load_i", code)
-        self.assertIn("work_tile_metadata.size(0)", code)
-        self.assertIn("'worklist_metadata': True", code)
-        self.assertIn("'real_groups_arg':", code)
-        self.assertNotIn("M_total_aligned + _BLOCK_SIZE_0 - 1", code)
-        self.assertIn("out = torch.zeros(", code)
-        self.assertNotIn("out = torch.empty(", code)
-
-        actual = mod.deepgemm_m_grouped_bf16_gemm_nt_contiguous(*args)
-        torch.cuda.synchronize()
-        torch.testing.assert_close(actual, expected, rtol=2e-2, atol=2e-2)
-        padding_rows = args[2] == -1
-        self.assertTrue(bool(torch.all(actual[padding_rows] == 0).item()))
-
-        graph = torch.cuda.CUDAGraph()
-        torch.cuda.synchronize()
-        with torch.cuda.graph(graph):
-            captured = mod.deepgemm_m_grouped_bf16_gemm_nt_contiguous(*args)
-        torch.cuda.synchronize()
-
-        captured[padding_rows] = 13
-        torch.cuda.synchronize()
-        graph.replay()
-        torch.cuda.synchronize()
-        torch.testing.assert_close(captured, expected, rtol=2e-2, atol=2e-2)
-        self.assertTrue(bool(torch.all(captured[padding_rows] == 0).item()))
-
-    @onlyBackends(["cute"])
-    @skipIfNotCUDA()
-    @skipIfRefEager("CUDA graph capture requires compiled kernel")
-    @patch.dict(
-        os.environ,
-        {
-            "HELION_BACKEND": "cute",
-            "HELION_CUTE_MMA_IMPL": "tcgen05",
-            "HELION_AUTOTUNE_EFFORT": "none",
-        },
-        clear=False,
-    )
-    def test_grouped_gemm_deepgemm_generated_segment_full_overwrite_uses_empty(
-        self,
-    ):
-        self._skip_unless_deepgemm_segment_tcgen05()
-        torch.manual_seed(0)
-        mod = import_path(EXAMPLES_DIR / "grouped_gemm.py")
-        args, expected = mod.make_deepgemm_m_grouped_bf16_gemm_nt_contiguous_args(
-            (188, 68),
-            n=128,
-            k=64,
-            m_alignment=1,
-            tail_padding=0,
-            device=DEVICE,
-        )
-        self.assertEqual(int(torch.count_nonzero(args[2] == -1).item()), 0)
-        layout_has_valid_prefix_tiles = (
-            mod._validate_deepgemm_m_grouped_bf16_gemm_nt_contiguous_args(*args)
-        )
-        self.assertFalse(layout_has_valid_prefix_tiles)
-        self.assertTrue(
-            mod._deepgemm_m_grouped_bf16_gemm_nt_contiguous_use_generated_segment_kernel(
-                *args,
-                layout_has_valid_prefix_tiles=layout_has_valid_prefix_tiles,
-            )
-        )
-        self.assertTrue(
-            mod._deepgemm_m_grouped_bf16_gemm_nt_contiguous_segment_output_fully_overwritten(
-                *args
-            )
-        )
-
-        work_tile_metadata = (
-            mod._deepgemm_m_grouped_bf16_gemm_nt_contiguous_segment_metadata_tensor(
-                *args
-            )
-        )
-        self.assertIsNotNone(work_tile_metadata)
-        self.assertEqual(
-            work_tile_metadata.detach().cpu().tolist(),
-            [
-                [0, 0, 128, 0],
-                [0, 128, 60, 0],
-                [1, 188, 68, 0],
-            ],
-        )
-        segment_kernel = mod._deepgemm_m_grouped_bf16_gemm_nt_contiguous_segment_full_overwrite_kernel
-        bound = segment_kernel.bind((args[0], args[1], work_tile_metadata))
-        bound.env.config_spec.cute_tcgen05_search_enabled = True
-        (config,) = segment_kernel.configs
-        code = bound.to_triton_code(config)
-        self.assertIn("cute.nvgpu.tcgen05", code)
-        self.assertIn("cute.gemm(", code)
-        self.assertIn("StaticPersistentGroupTileScheduler.create", code)
-        self.assertIn("out = torch.empty(", code)
-        self.assertNotIn("out = torch.zeros(", code)
-
-        actual = mod.deepgemm_m_grouped_bf16_gemm_nt_contiguous(*args)
-        torch.cuda.synchronize()
-        torch.testing.assert_close(actual, expected, rtol=2e-2, atol=2e-2)
-
-    @onlyBackends(["cute"])
-    @skipIfNotCUDA()
-    @patch.dict(
-        os.environ,
-        {
-            "HELION_BACKEND": "cute",
-            "HELION_CUTE_MMA_IMPL": "tcgen05",
-            "HELION_AUTOTUNE_EFFORT": "none",
-        },
-        clear=False,
-    )
-    def test_grouped_gemm_deepgemm_generated_segment_rejects_store_mismatch(
-        self,
-    ):
-        self._skip_unless_deepgemm_segment_tcgen05()
-        args, work_tile_metadata = self._deepgemm_segment_metadata_fixture()
-        for name, kernel in (
-            ("row", _deepgemm_segment_bad_store_row_kernel),
-            ("mask", _deepgemm_segment_bad_store_mask_kernel),
-            ("store_m", _deepgemm_segment_store_m_mask_kernel),
-        ):
-            with self.subTest(name=name):
-                self._assert_segment_tcgen05_not_admitted(
-                    kernel,
-                    args,
-                    work_tile_metadata,
-                )
-
-    @onlyBackends(["cute"])
-    @skipIfNotCUDA()
-    @patch.dict(
-        os.environ,
-        {
-            "HELION_BACKEND": "cute",
-            "HELION_CUTE_MMA_IMPL": "tcgen05",
-            "HELION_AUTOTUNE_EFFORT": "none",
-        },
-        clear=False,
-    )
-    def test_grouped_gemm_deepgemm_generated_segment_rejects_inverted_row_mask(
-        self,
-    ):
-        self._skip_unless_deepgemm_segment_tcgen05()
-        args, work_tile_metadata = self._deepgemm_segment_metadata_fixture()
-        self._assert_segment_tcgen05_not_admitted(
-            _deepgemm_segment_inverted_row_mask_kernel,
-            args,
-            work_tile_metadata,
-        )
-
-    @onlyBackends(["cute"])
-    @skipIfNotCUDA()
-    @patch.dict(
-        os.environ,
-        {
-            "HELION_BACKEND": "cute",
-        },
-        clear=False,
-    )
-    def test_grouped_gemm_deepgemm_generated_segment_cache_keys_preserve_capture_metadata(
-        self,
-    ):
-        mod = import_path(EXAMPLES_DIR / "grouped_gemm.py")
-        layout_validation_cache = mod._DEEPGEMM_LAYOUT_VALIDATION_CACHE
-        layout_segments_cache = mod._DEEPGEMM_CUTEDSL_LAYOUT_SEGMENTS_CACHE
-        segment_tensor_cache = mod._DEEPGEMM_CUTEDSL_SEGMENT_TENSOR_CACHE
-        selected_unsupported_cache = (
-            mod._DEEPGEMM_CUTEDSL_SELECTED_SEGMENT_UNSUPPORTED_CACHE
-        )
-        saved_layout_validation_cache = layout_validation_cache.copy()
-        saved_layout_segments_cache = layout_segments_cache.copy()
-        saved_segment_tensor_cache = segment_tensor_cache.copy()
-        saved_selected_unsupported_cache = selected_unsupported_cache.copy()
-        try:
-            layout_validation_cache.clear()
-            layout_segments_cache.clear()
-            segment_tensor_cache.clear()
-            selected_unsupported_cache.clear()
-
-            m_per_group = (188, 64, 40, 64)
-            layout: list[int] = []
-            for group_id, actual_m in enumerate(m_per_group):
-                layout.extend([group_id] * actual_m)
-                layout.extend([-1] * ((-actual_m) % 224))
-            args = (
-                torch.empty(len(layout), 64, device=DEVICE, dtype=torch.bfloat16),
-                torch.empty(
-                    len(m_per_group),
-                    128,
-                    64,
-                    device=DEVICE,
-                    dtype=torch.bfloat16,
-                ),
-                torch.tensor(layout, device=DEVICE, dtype=torch.int32),
-            )
-            legacy_metadata_key = mod._deepgemm_layout_validation_cache_key(
-                *args,
-                block_m=mod.DEEPGEMM_M_GROUPED_BF16_GEMM_NT_CONTIGUOUS_TILE_M,
-            )
-            segment_worklist_key = mod._deepgemm_layout_validation_cache_key(
-                *args,
-                block_m=mod.DEEPGEMM_M_GROUPED_BF16_GEMM_NT_CONTIGUOUS_SEGMENT_TILE_M,
-            )
-            selected_worklist_key = mod._deepgemm_layout_validation_cache_key(
-                *args,
-                block_m=mod.DEEPGEMM_M_GROUPED_BF16_GEMM_NT_CONTIGUOUS_SELECTED_SOURCE_M_TILE,
-            )
-            self.assertNotEqual(legacy_metadata_key, segment_worklist_key)
-            self.assertNotEqual(legacy_metadata_key, selected_worklist_key)
-            self.assertNotEqual(segment_worklist_key, selected_worklist_key)
-
-            layout_has_valid_prefix_tiles = (
-                mod._validate_deepgemm_m_grouped_bf16_gemm_nt_contiguous_args(*args)
-            )
-            self.assertFalse(layout_has_valid_prefix_tiles)
-            self.assertIsNotNone(
-                mod._deepgemm_m_grouped_bf16_gemm_nt_contiguous_cutedsl_metadata(*args)
-            )
-            work_tile_metadata = (
-                mod._deepgemm_m_grouped_bf16_gemm_nt_contiguous_segment_metadata_tensor(
-                    *args
-                )
-            )
-            self.assertIsNotNone(work_tile_metadata)
-            selected_work_tile_metadata = mod._deepgemm_m_grouped_bf16_gemm_nt_contiguous_selected_segment_metadata_tensor(
-                *args
-            )
-            self.assertIsNotNone(selected_work_tile_metadata)
-
-            self.assertIn(legacy_metadata_key, layout_segments_cache)
-            self.assertNotIn(segment_worklist_key, layout_segments_cache)
-            self.assertNotIn(selected_worklist_key, layout_segments_cache)
-            self.assertIn(segment_worklist_key, segment_tensor_cache)
-            self.assertIn(selected_worklist_key, segment_tensor_cache)
-            self.assertNotIn(legacy_metadata_key, segment_tensor_cache)
-
-            with (
-                patch.object(
-                    mod,
-                    "_parse_deepgemm_m_grouped_layout_segments",
-                    side_effect=AssertionError(
-                        "unexpected layout parse during capture"
-                    ),
-                ),
-                patch.object(
-                    mod,
-                    "_deepgemm_cuda_graph_capture_active",
-                    return_value=True,
-                ),
-            ):
-                self.assertFalse(
-                    mod._validate_deepgemm_m_grouped_bf16_gemm_nt_contiguous_args(*args)
-                )
-                self.assertIsNotNone(
-                    mod._deepgemm_m_grouped_bf16_gemm_nt_contiguous_cutedsl_metadata(
-                        *args
-                    )
-                )
-                self.assertTrue(
-                    mod._deepgemm_m_grouped_bf16_gemm_nt_contiguous_use_generated_segment_kernel(
-                        *args,
-                        layout_has_valid_prefix_tiles=layout_has_valid_prefix_tiles,
-                    )
-                )
-                self.assertIs(
-                    mod._deepgemm_m_grouped_bf16_gemm_nt_contiguous_segment_metadata_tensor(
-                        *args
-                    ),
-                    work_tile_metadata,
-                )
-                self.assertIs(
-                    mod._deepgemm_m_grouped_bf16_gemm_nt_contiguous_selected_segment_metadata_tensor(
-                        *args
-                    ),
-                    selected_work_tile_metadata,
-                )
-        finally:
-            layout_validation_cache.clear()
-            layout_validation_cache.update(saved_layout_validation_cache)
-            layout_segments_cache.clear()
-            layout_segments_cache.update(saved_layout_segments_cache)
-            segment_tensor_cache.clear()
-            segment_tensor_cache.update(saved_segment_tensor_cache)
-            selected_unsupported_cache.clear()
-            selected_unsupported_cache.update(saved_selected_unsupported_cache)
-
-    @onlyBackends(["cute"])
-    @skipIfNotCUDA()
-    @patch.dict(
-        os.environ,
-        {
-            "HELION_BACKEND": "cute",
-        },
-        clear=False,
-    )
-    def test_grouped_gemm_deepgemm_valid_prefix_uses_legacy_generated_kernel(
-        self,
-    ):
-        torch.manual_seed(0)
-        mod = import_path(EXAMPLES_DIR / "grouped_gemm.py")
-        args, expected = mod.make_deepgemm_m_grouped_bf16_gemm_nt_contiguous_args(
-            (64, 96, 128, 32),
-            n=64,
-            k=64,
-            m_alignment=64,
-            tail_padding=64,
-            device=DEVICE,
-        )
-        layout_has_valid_prefix_tiles = (
-            mod._validate_deepgemm_m_grouped_bf16_gemm_nt_contiguous_args(*args)
-        )
-        self.assertTrue(layout_has_valid_prefix_tiles)
-        self.assertFalse(
-            mod._deepgemm_m_grouped_bf16_gemm_nt_contiguous_use_generated_segment_kernel(
-                *args,
-                layout_has_valid_prefix_tiles=layout_has_valid_prefix_tiles,
-            )
-        )
-
-        with (
-            patch.object(
-                mod,
-                "_deepgemm_m_grouped_bf16_gemm_nt_contiguous_generated_segment",
-                side_effect=AssertionError("segment fallback must not run"),
-            ),
-            patch.object(
-                mod,
-                "_deepgemm_m_grouped_bf16_gemm_nt_contiguous_kernel",
-                return_value=expected,
-            ) as legacy_kernel,
-        ):
-            actual = mod.deepgemm_m_grouped_bf16_gemm_nt_contiguous(*args)
-        legacy_kernel.assert_called_once_with(*args)
-        self.assertIs(actual, expected)
-
-    @onlyBackends(["cute"])
-    def test_grouped_gemm_deepgemm_m_grouped_layout_parser_rejects_invalid(self):
-        mod = import_path(EXAMPLES_DIR / "grouped_gemm.py")
-
-        def validate_layout(layout: list[int]) -> None:
-            rows = len(layout)
-            a_packed = torch.empty(rows, 8, device=DEVICE, dtype=torch.bfloat16)
-            b_grouped = torch.empty(4, 8, 8, device=DEVICE, dtype=torch.bfloat16)
-            grouped_layout = torch.tensor(layout, device=DEVICE, dtype=torch.int32)
-            mod._validate_deepgemm_m_grouped_bf16_gemm_nt_contiguous_args(
-                a_packed,
-                b_grouped,
-                grouped_layout,
-                block_m=4,
-            )
-
-        invalid_layouts = (
-            ([0, 0, 2, 2, 1, 1], "strictly increasing"),
-            ([0, 0, -1, 1, 1, -1, 1, 1], "strictly increasing"),
-            ([0, 0, -1, 0], "strictly increasing"),
-            ([0, -2, 1], "values must be -1"),
-        )
-        for layout, error in invalid_layouts:
-            with (
-                self.subTest(layout=layout),
-                self.assertRaisesRegex(ValueError, error),
-            ):
-                validate_layout(layout)
 
     def test_geglu(self):
         args = (
